Added legal notice Link for the frontpage
In Germany (and maybe some other countrys too) almost every website needs a legal notice which should be accessable from the frontpage.
I added the legalnotice with the same functionality like the CV. So if you don't set the legal notice there will be no changes to the website at all.